separating gossiper
